(index) Optimize compiled query evaluation
Originally the compiled queries were evaluated using various tree-traversal methods that have poor cache locality and somewhat sub-par performance.

This is replaced by an amortized mapping out of each path through the query, that is re-used for each result during ranking, so that evaluation becomes as easy as just walking through some flat arrays.

If the queries were even more complex, a stack machine based solution may have been a decent alternative, though the de-facto real world complexity of the expression trees are simply enough that this is not useful.