(index) Reject index executions when the index is overloaded
This avoids spinning up a lot of parallel executions, which would interfere with each other, slow the system, and perpetuate the problem.  This has been observed mainly during the startup phase, when caches and cold and JIT hasn't gotten very far.

A new system property 'index.maxSimultaneousQueries' is added as a tuning parameter.  The default value of 4 is probably on the high end.